facenet/requirements.txt
deepface>=0.0.90
pgvector>=0.4.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
requests>=2.31.0
tensorflow>=2.15.0
numpy>=1.24.0
opencv-python>=4.5.5
retina-face>=0.0.17
